+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-07-10
+
 ### Added
 
+- Obligatory `ruthless.java.languageVersion` system property, which configures Java toolchain language version
+- Java 8 compatibility
+- Declare Javadoc and Source JARs
 - Add Mockito BOM 4.6.1
 - Add `mockito-core` and `mockito-junit-jupiter` as default test dependencies
 - Add Ruthless Logging with capability to record build output to file
@@ -61,7 +66,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `dev.lajoscseppento.ruthless.spring-boot-library` - Spring Boot library
 - Publication to Gradle Plugin Portal
 
-[Unreleased]: https://github.com/LajosCseppento/ruthless/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/LajosCseppento/ruthless/compare/v0.4.0...HEAD
+
+[0.4.0]: https://github.com/LajosCseppento/ruthless/releases/tag/v0.4.0
 
 [0.3.0]: https://github.com/LajosCseppento/ruthless/releases/tag/v0.3.0